% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contrasts-package.R
\docType{package}
\name{Contrasts-package}
\alias{Contrasts}
\alias{Contrasts-package}
\title{Contrasts: Weighted sum-to-zero contrasts}
\description{
This contains some functions to convert a factor with k levels into a set of k-1 orthogonal sum-to-zero contrasts that are weighted by the inverse sample sizes of the groups. It does not assign contrasts to a factor in the traditional way. There is also a function for creating contrasts for a factor y nested in x.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anhsmith.github.io/Contrasts/}
}

}
\author{
\strong{Maintainer}: Adam Smith \email{anhsmith@gmail.com} (\href{https://orcid.org/0000-0003-0059-6206}{ORCID})

Authors:
\itemize{
  \item Marti Anderson \email{marti@primer-e.co.nz} (\href{https://orcid.org/0000-0002-4018-4049}{ORCID})
}

}
\keyword{internal}
