% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_weightzero_contrasts.R
\name{make_weightzero_contrasts}
\alias{make_weightzero_contrasts}
\title{Make weighted sum-to-zero contrasts for a factor}
\usage{
make_weightzero_contrasts(x)
}
\arguments{
\item{x}{}
}
\value{
cont
}
\description{
Make weighted sum-to-zero contrasts for a factor
}
\examples{
x <- factor(rep(letters[1:3], times=4:6))
make_weightzero_contrasts(x)
}
