% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_weightzero_nested_contrasts.R
\name{make_weightzero_nested_contrasts}
\alias{make_weightzero_nested_contrasts}
\title{Make weighted sum-to-zero contrasts for a factor y_in_x nested in x}
\usage{
make_weightzero_nested_contrasts(y_in_x, x)
}
\arguments{
\item{y_in_x}{A factor that is nested in x}

\item{x}{A factor in which y_in_x is nested}
}
\value{
\code{make_weightzero_nested_contrasts}
}
\description{
Make weighted sum-to-zero contrasts for a factor y_in_x nested in x
}
\examples{
x = factor(rep(letters[1:3], times=4:6))
y_in_x = factor(paste0(x,c(1,1,2,2,1,1,2,2,3,1,2,2,2,3,3)))

dplyr::bind_cols(
  x=x, y_in_x=y_in_x,
  make_weightzero_contrasts(x),
  make_weightzero_nested_contrasts(y_in_x, x)
)
}
